LayoutTests/http/tests/preload/dynamic_remove_preload_href.html

<!DOCTYPE html>
<html>
<head>
<script>
    if (window.internals)
        window.internals.setLinkPreloadSupport(true);
</script>
<script src="/js-test-resources/js-test.js"></script>
<body>
<script>
    var link = document.createElement("link");
    link.as = "script";
    link.rel = "preload";
    link.href = "../resources/dummy.js";
    document.body.appendChild(link);
    link.href = "";
    window.addEventListener("load", function() {
        shouldBeFalse("internals.isPreloaded('../resources/dummy.js');");
        if (window.internals)
            window.internals.setLinkPreloadSupport(false);
    });
</script>
</body>
